requirements.txt
amqp 
appdirs
atomicwrites
attrs
beautifulsoup4
certifi
chardet
codeforces-api
colorama
coverage
cycler
decorator 
django-crispy-forms
django-popup-view-field 
django-registration-redux
evtxtract 
hexdump
idna 
ipython-genutils 
Jinja2  
jsonschema
jupyter-core  
kiwisolver 
lxml   
MarkupSafe   
matplotlib    
more-itertools  
mysqlclient
nbformat 
numpy
packaging   
pbr   
plotly   
pluggy   
py    
pyparsing  
pytest   
pytest-cov 
python-dateutil   
python-evtx 
pytz   
requests 
retrying 
schedule  
scipy    
setuptools  
six  
stevedore
traitlets
urllib3  
vine  